Refactor command handling into its own module
It currently does quite a bit more than command handling, but its a
start for modularization.
